% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_examples.R
\name{export_examples}
\alias{export_examples}
\title{Export built-in examples to a user-specified folder.}
\usage{
export_examples(path_workingFolder)
}
\arguments{
\item{path_workingFolder}{user-specified folder path. Must be a valid path to a local folder.}
}
\value{
empty, writing to files only
}
\description{
This function takes in a user-specified folder and copy-paste the built-in examples (TLR shell, dummy TLFs, section structure csv, etc.) into the folder
}
\examples{
# By default, not returning to file, pass the officer object to the next function one by one
export_examples(path_workingFolder = "C:/Documents/TLR_folder/")

}
