% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_demo.R
\name{import_demo}
\alias{import_demo}
\title{Create a copy of demo_TLR.R within the current environment}
\usage{
import_demo()
}
\description{
This function copy the example code and create a copy called "demo_TLR.R" in the current working directory
}
\examples{
# Just run the import_demo() function to create a copy called "demo_TLR.R" in your current working environment
import_demo( )

}
