% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_RTFs.R
\name{import_RTFs}
\alias{import_RTFs}
\title{Create a copy of RTFs outputs within the current environment}
\usage{
import_RTFs()
}
\description{
This function copy the dummy RTF outputs from the package and create a folder called "rtf_outputs" in the current working directory
}
\examples{
# Just run the import_RTFs() function to create a copy folder within the RTFs outputs in your current working environment
RapTLR::import_RTFs( )

}
