% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_TLFs.R
\name{import_TLFs}
\alias{import_TLFs}
\title{Create a copy of TLFs folder within the current environment}
\usage{
import_TLFs()
}
\description{
This function copy the TLFs folder from RapTLR and create a copy called "TLF_outputs" in the current working directory
}
\examples{
# Just run the import_TLFs() function to create a copy folder TLF_outputs in your current working environment
RapTLR::import_TLFs( )

}
