% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apdx.R
\name{run_apdx}
\alias{run_apdx}
\title{Create the appendix of a docx file - combine link creation and add TLF tables}
\usage{
run_apdx(
  path_TLFs,
  docx_object,
  keyword = "LISTOFAPPENDICES",
  sections_structure = NULL,
  return_to_file = NULL
)
}
\arguments{
\item{path_TLFs}{Path to the folder where your table outputs (TLFs) are saved.}

\item{docx_object}{Path to the folder where your target docx is saved.}

\item{keyword}{IMPORTANT: please ensure your doc has this unique keyword placed where you want the Appendix section to start.}

\item{return_to_file}{Optionnal: Path and name of the docx document with the appendice. If NULL then the new docx document will be saved at the working directory as appendix_processed.docx}

\item{sections_structure:}{Optionnal: If NULL then all outputs from path_TLFs will be used and no section will be created. Otherwise format of list, csv or excel and subsections.}
}
\value{
NULL, output a docx to the destination folder.
}
\description{
This function takes an empty (or headed) docx file, links your output TLF files, and adds them to the docx. Please note that this function will automatically create a new folder "new_linked" within your path_TLFs for processing intermediate TLF files.
}
\examples{
path_TLFs <- system.file("extdata/TLF_outputs", package = "RapTLR")
path_docx <- system.file("extdata", "TLR_shell.docx", package = "RapTLR")
path_result <- file.path(getwd( ), "docx_with_appendice")
# By default, no sections are specified. All tables are automatically grouped into a single section called "Outputs" in the appendice and saved in the current working directory as appendix_processed.docx.
run_apdx(path_TLFs = path_TLFs,
         docx_object = file.path(path_docx, "TLR_shell.docx"))

# The path and name of the new docx object can be modifed with return_to_file argument.
run_apdx(path_TLFs = path_TLFs,
         docx_object = path_docx,
         return_to_file = path_result)

# A CSV file can be used to define the structure of the appendices.
TLF_list_csv <- system.file("extdata", "TLF_list.csv", package = "RapTLR")

run_apdx(path_TLFs = path_TLFs,
         docx_object = path_docx,
         sections_structure = TLF_list_csv,
         return_to_file = path_result)

# Or an excel file 
TLF_list_xlsx <- system.file("extdata", "TLF_list.xlsx", package = "RapTLR")

run_apdx(path_TLFs = path_TLFs,
         docx_object = path_docx,
         sections_structure = TLF_list_xlsx,
         return_to_file = path_result)

# Or a list following the below format :
output_list <- list(Efficacy = "tefmad01a",
                    Overall = "tsidem03",
                    Safety = c("tsfae10", "lsfae03"))
run_apdx(path_TLFs = path_TLFs,
         docx_object = path_docx,
         sections_structure = output_list)


}
