% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apdx.R
\name{run_apdx}
\alias{run_apdx}
\title{Create the appendix of a docx file - combine link creation and add TLF tables}
\usage{
run_apdx(
  path_TLFs,
  docx_object,
  keyword = "LISTOFAPPENDICES",
  sections_structure = NULL,
  return_to_file = NULL,
  return_object = FALSE
)
}
\arguments{
\item{path_TLFs}{Path to the folder where your table outputs (TLFs) are saved.}

\item{docx_object}{Path to the folder where your target docx is saved.}

\item{keyword}{IMPORTANT: please ensure your doc has this unique keyword placed where you want the Appendix section to start.}

\item{return_to_file}{Optional - if not specified, saving result docx file to the same folder with default name "RapTLR_Appendix_processed.docx". This function can auto-detect if the user specified a docx name along with it. If not, auto-create a file name and save to file.}

\item{return_object}{Optional - if set to T, the function will return an officer object in global environment regardless of return_to_file}

\item{output}{FOR LATER: To select only some desired TLFs and not all.}

\item{sections_structure:}{If set to Null then all outputs from path_TLFs will be used and no section will be created. Otherwise format of list, csv or excel and subsections.}
}
\value{
NULL, output a docx to the destination folder.
}
\description{
This function takes an empty (or headed) docx file, links your output TLF files, and adds them to the docx. Please note that this function will automatically create a new folder "new_linked" within your path_TLFs for processing intermediate TLF files.
}
\examples{
path_TLFs = "REPLACE WITH YOUR OWN PATH C:/Documents/TLR_folder/TLF_outputs"
path_docx = "REPLACE WITH YOUR OWN PATH C:/Documents/TLR_folder/"
path_result = "REPLACE WITH YOUR OWN PATH C:/Documents/TLR_folder/TLR_shell_processed.docx"
# By default, no sections are specified. All tables are automatically grouped into a single section called "Outputs" in the appendice.
run_apdx(path_TLFs = path_TLFs,
          docx_object = file.path(path_docx, "TLR_shell.docx"))

# An excel file can be read to define the structure of the appendice with the corresponding outputs.
run_apdx(path_TLFs = path_TLFs,
         docx_object = file.path(path_docx, "TLR_shell.docx"),
         keyword = "LISTOFAPPENDICES",
         return_to_file = path_result)

# A CSV file can also be used to define the structure of the appendices.
run_apdx(path_TLFs = path_TLFs,
         docx_object = file.path(path_docx, "TLR_shell.docx"),
         keyword = "LISTOFAPPENDICES",
         sections_structure = "REPLACE WITH YOUR OWN PATH C:/Documents/TLR_folder/TLF_list.csv",
         return_to_file = path_result)

# Or a list following the below format :
output_list <- list(Efficacy = "tefmad01a",
                    Overall = "tsidem03",
                    Safety = c("tsfae10", "lsfae03"))
run_apdx(path_TLFs = path_TLFs,
         docx_object = file.path(path_docx, "TLR_shell.docx"),
         sections_structure = output_list)


}
