% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textReplace.R
\name{textReplace}
\alias{textReplace}
\title{Detect texts marked by keywords in the doc and replace them with dynamic text & numbers.}
\usage{
textReplace(doc_object, keyword, replacement, return_to_file = NULL)
}
\arguments{
\item{doc_object}{Can be either a string or officer object. If string, a path to the folder where your doc or docx is saved. If officer object, should be the doc/docx loaded by officer::read_docx()}

\item{keyword}{Keyword in the document to be replaced.}

\item{replacement}{The text that you want to insert at the keyword position.}

\item{return_to_file}{Optional path to save the modified document. If NULL, returns officer object.}
}
\value{
officer object or writing to files instead
}
\description{
This function takes a docx template file (e.g. TLR shell with keywords like "TTsubjectTT"), and replace them with texts dynamically generated by user's R code.
}
\examples{
# You can provide docx object
docx_path <- system.file("extdata", "Keywords_replace.docx", package = "RapTLR")

newdocx <- textReplace(doc_object = docx_path,
                       keyword = "TTsubjectTT",
                       replacement = "This is new text")

# You can also provide docx object in the environment 
docx_path <- system.file("extdata", "Keywords_replace.docx", package = "RapTLR")
docx_imported <- officer::read_docx(docx_path)

newdocx <- textReplace(doc_object = docx_imported,
                       keyword = "TTsubjectTT",
                       replacement = "This is new text")

# You can provide the name of the new document
docx_imported <- officer::read_docx(docx_path)
textReplace(doc_object = docx_imported,
            keyword = "TTsubjectTT",
            replacement = "This is new text",
            return_to_file = file.path( getwd(), "new_docx.docx" ))

# You can use R's built-in or dplyr's function chains to chain your functions
docx_imported <- officer::read_docx(docx_path)
textReplace(doc_object = docx_imported,
            keyword = "TTsubjectTT",
            replacement = "This is new text")  |>
textReplace(keyword = "TTsiteTT", 
            replacement = "This is new text 2") |>
textReplace(keyword = "TTconclusionTT", 
            replacement = "This is new text 3", 
            return_to_file = "TLR_shell_replaced_texts.docx")

# See help(run_apdx) or Github landing page for additional details..
}
